requirements.txt
modelcatalog-api
typing
obasparql==2.0.2
Click
